// frontend/src/app/api-gen/models/diff-response.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface DiffResponse {
  code_a?: (string | null);
  code_b?: (string | null);
  env_a: string;
  env_b: string;
  notes?: Array<string>;
  object_name: string;
  object_type: string;
  schema_name: string;
  script?: (string | null);
  status: string;
}